research gate

+++ resources/views/admin/news/index.blade.php
@@ -1,6 +1,7 @@
 @extends('layouts.template_demo')
 
 @section('content')
+
 <div class="right_col" role="main">
     <div class="page-title">
         <div class="title_left"><a href="{{ route('news.index') }}"><h3>News</h3></a></div>
@@ -37,23 +38,25 @@
                 </thead>
                 <tbody>
                     @foreach ($news as $article)
-                        <tr>
-                            <td><a href="{{ route('news.show', $article ) }}"><small>{{ $article->title }}</small></a></td>
-                            <td><small>{{ $article->category ? $article->category->name : '' }}</small></td>
-                            <td>
-                                <label style="text-overflow: ellipsis; height:75px; width:200px; border:1px solid; padding:2px 5px; overflow:hidden">
-                                    <small>{!! $article->summary !!}</small>
-                                </label>
-                            </td>
-                            <td><img src="{{ asset('images/bai_viet/' . $article->image ) }}" alt="news" style="width:70px; height:70px"></td>
-                            <td>
-                                {{ Form::open(['route' => ['news.destroy',$article->id], 'method' => 'delete']) }}
-                                    {{-- <a href="{{ route('news.show', $article) }}" class="btn btn-primary btn-xs"><i class="fa fa-folder"></i></a> --}}
-                                    <a href="{{ route('news.edit', $article) }}" class="btn btn-info btn-xs"><i class="fa fa-pencil"></i></a>
-                                    {{-- {!! Form::button('<i class="fa fa-trash-o"></i>', ['class'=>'btn btn-danger btn-xs', 'type'=>'submit']) !!} --}}
-                                {{ Form::close() }}
-                            </td>
-                        </tr>
+                        @can('index', $article)
+                            <tr>
+                                <td><a href="{{ route('news.show', $article ) }}"><small>{{ $article->title }}</small></a></td>
+                                <td><small>{{ $article->category ? $article->category->name : '' }}</small></td>
+                                <td>
+                                    <label style="text-overflow: ellipsis; height:75px; width:200px; border:1px solid; padding:2px 5px; overflow:hidden">
+                                        <small>{!! $article->summary !!}</small>
+                                    </label>
+                                </td>
+                                <td><img src="{{ asset('images/bai_viet/' . $article->image ) }}" alt="news" style="width:70px; height:70px"></td>
+                                <td>
+                                    {{ Form::open(['route' => ['news.destroy',$article->id], 'method' => 'delete']) }}
+                                        {{-- <a href="{{ route('news.show', $article) }}" class="btn btn-primary btn-xs"><i class="fa fa-folder"></i></a> --}}
+                                        <a href="{{ route('news.edit', $article) }}" class="btn btn-info btn-xs"><i class="fa fa-pencil"></i></a>
+                                        {{-- {!! Form::button('<i class="fa fa-trash-o"></i>', ['class'=>'btn btn-danger btn-xs', 'type'=>'submit']) !!} --}}
+                                    {{ Form::close() }}
+                                </td>
+                            </tr>
+                        @endcan
                     @endforeach
                 </tbody>
             </table>
@@ -61,4 +64,5 @@
         </div>
     </div>
 </div>
+
 @endsection
